Clone-happy removal of Copy from UpvarId
We would like to extend UpvarId to also carry an access
path (i.e. Upvar (x, .y) for an access to x.y). However, UpvarIds are
used as keys all over the place. This patch tries to remove the need
for UpvarId to be Copy, by liberally cloning it wherever needed. I
doubt this has performance implications but OTOH it goes against the
"put it in a table and use the index as the ID" practice all over the
compiler.