Propagate captures to parent paths
When we capture x.y as immutable but x.y.z as mutable, that forces us
to also capture x.y as mutable. Not sure if this is minimal enough;
technically, we could perhaps try to break up x.y and capture everything
other than x.y.z as immutable. Discuss.